+++ app/scripts/background.js
@@ -448,6 +448,7 @@ function triggerUi () {
     const currentlyActiveMetamaskTab = Boolean(tabs.find(tab => openMetamaskTabsIDs[tab.id]))
     if (!popupIsOpen && !currentlyActiveMetamaskTab && !notificationIsOpen) {
       notificationManager.showPopup()
+      notificationIsOpen = true
     }
   })
 }