+++ test/e2e/beta/metamask-beta-responsive-ui.spec.js
@@ -1,7 +1,7 @@
 const path = require('path')
 const assert = require('assert')
 const webdriver = require('selenium-webdriver')
-const { By, Key, until } = webdriver
+const { By, until } = webdriver
 const {
   delay,
   buildChromeWebDriver,
@@ -13,22 +13,19 @@ const {
 const {
   checkBrowserForConsoleErrors,
   closeAllWindowHandlesExcept,
-  verboseReportOnFailure,
   findElement,
   findElements,
   loadExtension,
+  verboseReportOnFailure,
 } = require('./helpers')
 
-
-describe('Using MetaMask with an existing account', function () {
+describe('MetaMask', function () {
   let extensionId
   let driver
 
   const testSeedPhrase = 'phrase upgrade clock rough situate wedding elder clever doctor stamp excess tent'
-  const testAddress = '0xE18035BF8712672935FDB4e5e431b1a0183d2DFC'
-  const testPrivateKey2 = '14abe6f4aab7f9f626fe981c864d0adeb5685f289ac9270c27b8fd790b4235d6'
-  const tinyDelayMs = 500
-  const regularDelayMs = 1000
+  const tinyDelayMs = 200
+  const regularDelayMs = tinyDelayMs * 2
   const largeDelayMs = regularDelayMs * 2
 
   this.timeout(0)
@@ -37,22 +34,19 @@ describe('Using MetaMask with an existing account', function () {
   before(async function () {
     switch (process.env.SELENIUM_BROWSER) {
       case 'chrome': {
-        const extensionPath = path.resolve('dist/chrome')
-        driver = buildChromeWebDriver(extensionPath)
+        const extPath = path.resolve('dist/chrome')
+        driver = buildChromeWebDriver(extPath, { responsive: true })
         extensionId = await getExtensionIdChrome(driver)
         await driver.get(`chrome-extension://${extensionId}/popup.html`)
-        await delay(regularDelayMs)
         break
       }
       case 'firefox': {
-        const extensionPath = path.resolve('dist/firefox')
-        driver = buildFirefoxWebdriver()
-        await installWebExt(driver, extensionPath)
-        await delay(regularDelayMs)
+        const extPath = path.resolve('dist/firefox')
+        driver = buildFirefoxWebdriver({ responsive: true })
+        await installWebExt(driver, extPath)
+        await delay(700)
         extensionId = await getExtensionIdFirefox(driver)
         await driver.get(`moz-extension://${extensionId}/popup.html`)
-        await delay(regularDelayMs)
-        break
       }
     }
   })
@@ -137,35 +131,33 @@ describe('Using MetaMask with an existing account', function () {
     })
   })
 
-  describe('First time flow starting from an existing seed phrase', () => {
-    it('imports a seed phrase', async () => {
-      const [seedPhrase] = await findElements(driver, By.xpath(`//a[contains(text(), 'Import with seed phrase')]`))
-      await seedPhrase.click()
+  describe('Going through the first time flow', () => {
+    it('accepts a secure password', async () => {
+      const passwordBox = await findElement(driver, By.css('.create-password #create-password'))
+      const passwordBoxConfirm = await findElement(driver, By.css('.create-password #confirm-password'))
+      const button = await findElement(driver, By.css('.create-password button'))
+
+      await passwordBox.sendKeys('correct horse battery staple')
+      await passwordBoxConfirm.sendKeys('correct horse battery staple')
+      await button.click()
       await delay(regularDelayMs)
+    })
 
-      const [seedTextArea] = await findElements(driver, By.css('textarea.import-account__secret-phrase'))
-      await seedTextArea.sendKeys(testSeedPhrase)
-      await delay(regularDelayMs)
-
-      const [password] = await findElements(driver, By.id('password'))
-      await password.sendKeys('correct horse battery staple')
-      const [confirmPassword] = await findElements(driver, By.id('confirm-password'))
-      confirmPassword.sendKeys('correct horse battery staple')
-
-      const [importButton] = await findElements(driver, By.xpath(`//button[contains(text(), 'Import')]`))
-      await importButton.click()
+    it('clicks through the unique image screen', async () => {
+      const nextScreen = await findElement(driver, By.css('.unique-image button'))
+      await nextScreen.click()
       await delay(regularDelayMs)
     })
 
     it('clicks through the ToS', async () => {
       // terms of use
-      await delay(largeDelayMs)
       const canClickThrough = await driver.findElement(By.css('.tou button')).isEnabled()
       assert.equal(canClickThrough, false, 'disabled continue button')
       const bottomOfTos = await findElement(driver, By.linkText('Attributions'))
       await driver.executeScript('arguments[0].scrollIntoView(true)', bottomOfTos)
       await delay(regularDelayMs)
       const acceptTos = await findElement(driver, By.css('.tou button'))
+      driver.wait(until.elementIsEnabled(acceptTos))
       await acceptTos.click()
       await delay(regularDelayMs)
     })
@@ -186,53 +178,120 @@ describe('Using MetaMask with an existing account', function () {
       await nextScreen.click()
       await delay(regularDelayMs)
     })
+
+    let seedPhrase
+
+    it('reveals the seed phrase', async () => {
+      const byRevealButton = By.css('.backup-phrase__secret-blocker .backup-phrase__reveal-button')
+      await driver.wait(until.elementLocated(byRevealButton, 10000))
+      const revealSeedPhraseButton = await findElement(driver, byRevealButton, 10000)
+      await revealSeedPhraseButton.click()
+      await delay(regularDelayMs)
+
+      seedPhrase = await driver.findElement(By.css('.backup-phrase__secret-words')).getText()
+      assert.equal(seedPhrase.split(' ').length, 12)
+      await delay(regularDelayMs)
+
+      const nextScreen = await findElement(driver, By.css('.backup-phrase button'))
+      await nextScreen.click()
+      await delay(regularDelayMs)
+    })
+
+    async function clickWordAndWait (word) {
+      const xpathClass = 'backup-phrase__confirm-seed-option backup-phrase__confirm-seed-option--unselected'
+      const xpath = `//button[@class='${xpathClass}' and contains(text(), '${word}')]`
+      const word0 = await findElement(driver, By.xpath(xpath), 10000)
+
+      await word0.click()
+      await delay(tinyDelayMs)
+    }
+
+    async function retypeSeedPhrase (words, wasReloaded, count = 0) {
+      try {
+        if (wasReloaded) {
+          const byRevealButton = By.css('.backup-phrase__secret-blocker .backup-phrase__reveal-button')
+          await driver.wait(until.elementLocated(byRevealButton, 10000))
+          const revealSeedPhraseButton = await findElement(driver, byRevealButton, 10000)
+          await revealSeedPhraseButton.click()
+          await delay(regularDelayMs)
+
+          const nextScreen = await findElement(driver, By.css('.backup-phrase button'))
+          await nextScreen.click()
+          await delay(regularDelayMs)
+        }
+
+        for (let i = 0; i < 12; i++) {
+          await clickWordAndWait(words[i])
+        }
+      } catch (e) {
+        if (count > 2) {
+          throw e
+        } else {
+          await loadExtension(driver, extensionId)
+          await retypeSeedPhrase(words, true, count + 1)
+        }
+      }
+    }
+
+    it('can retype the seed phrase', async () => {
+      const words = seedPhrase.split(' ')
+
+      await retypeSeedPhrase(words)
+
+      const confirm = await findElement(driver, By.xpath(`//button[contains(text(), 'Confirm')]`))
+      await confirm.click()
+      await delay(regularDelayMs)
+    })
+
+    it('clicks through the deposit modal', async () => {
+      const byBuyModal = By.css('span .modal')
+      const buyModal = await driver.wait(until.elementLocated(byBuyModal))
+      const closeModal = await findElement(driver, By.css('.page-container__header-close'))
+      await closeModal.click()
+      await driver.wait(until.stalenessOf(buyModal))
+      await delay(regularDelayMs)
+    })
   })
 
   describe('Show account information', () => {
-    it('shows the correct account address', async () => {
-      await driver.findElement(By.css('.wallet-view__details-button')).click()
-      await driver.findElement(By.css('.qr-wrapper')).isDisplayed()
-      await delay(regularDelayMs)
-
-      const [address] = await findElements(driver, By.css('input.qr-ellip-address'))
-      assert.equal(await address.getAttribute('value'), testAddress)
-
-      await driver.executeScript("document.querySelector('.account-modal-close').click()")
-      await delay(largeDelayMs)
-    })
-
-    it('shows a QR code for the account', async () => {
-      await driver.findElement(By.css('.wallet-view__details-button')).click()
-      await driver.findElement(By.css('.qr-wrapper')).isDisplayed()
-      const detailModal = await driver.findElement(By.css('span .modal'))
-      await delay(regularDelayMs)
-
-      await driver.executeScript("document.querySelector('.account-modal-close').click()")
-      await driver.wait(until.stalenessOf(detailModal))
+    it('show account details dropdown menu', async () => {
+      await driver.findElement(By.css('div.menu-bar__open-in-browser')).click()
+      const options = await driver.findElements(By.css('div.menu.account-details-dropdown div.menu__item'))
+      assert.equal(options.length, 3) // HD Wallet type does not have to show the Remove Account option
       await delay(regularDelayMs)
     })
   })
 
-  describe('Log out and log back in', () => {
-    it('logs out of the account', async () => {
-      const accountIdenticon = driver.findElement(By.css('.account-menu__icon .identicon'))
-      accountIdenticon.click()
+  describe('Import seed phrase', () => {
+    it('logs out of the vault', async () => {
+      await driver.findElement(By.css('.account-menu__icon')).click()
       await delay(regularDelayMs)
 
-      const [logoutButton] = await findElements(driver, By.css('.account-menu__logout-button'))
+      const logoutButton = await findElement(driver, By.css('.account-menu__logout-button'))
       assert.equal(await logoutButton.getText(), 'Log out')
       await logoutButton.click()
       await delay(regularDelayMs)
     })
 
-    it('accepts the account password after lock', async () => {
-      await driver.findElement(By.id('password')).sendKeys('correct horse battery staple')
-      await driver.findElement(By.id('password')).sendKeys(Key.ENTER)
-      await delay(largeDelayMs)
-    })
-  })
+    it('imports seed phrase', async () => {
+      const restoreSeedLink = await findElement(driver, By.css('.unlock-page__link--import'))
+      assert.equal(await restoreSeedLink.getText(), 'Import using account seed phrase')
+      await restoreSeedLink.click()
+      await delay(regularDelayMs)
 
-  describe('Add an account', () => {
+      const seedTextArea = await findElement(driver, By.css('textarea'))
+      await seedTextArea.sendKeys(testSeedPhrase)
+      await delay(regularDelayMs)
+
+      const passwordInputs = await driver.findElements(By.css('input'))
+      await delay(regularDelayMs)
+
+      await passwordInputs[0].sendKeys('correct horse battery staple')
+      await passwordInputs[1].sendKeys('correct horse battery staple')
+      await driver.findElement(By.css('.first-time-flow__button')).click()
+      await delay(regularDelayMs)
+    })
+
     it('switches to localhost', async () => {
       const networkDropdown = await findElement(driver, By.css('.network-name'))
       await networkDropdown.click()
@@ -240,42 +299,12 @@ describe('Using MetaMask with an existing account', function () {
 
       const [localhost] = await findElements(driver, By.xpath(`//span[contains(text(), 'Localhost')]`))
       await localhost.click()
-      await delay(largeDelayMs)
+      await delay(largeDelayMs * 2)
     })
 
-    it('choose Create Account from the account menu', async () => {
-      await driver.findElement(By.css('.account-menu__icon')).click()
-      await delay(regularDelayMs)
-
-      const [createAccount] = await findElements(driver, By.xpath(`//div[contains(text(), 'Create Account')]`))
-      await createAccount.click()
-      await delay(regularDelayMs)
-    })
-
-    it('set account name', async () => {
-      const [accountName] = await findElements(driver, By.css('.new-account-create-form input'))
-      await accountName.sendKeys('2nd account')
-      await delay(regularDelayMs)
-
-      const [createButton] = await findElements(driver, By.xpath(`//button[contains(text(), 'Create')]`))
-      await createButton.click()
-      await delay(regularDelayMs)
-    })
-
-    it('should show the correct account name', async () => {
-      const [accountName] = await findElements(driver, By.css('.account-name'))
-      assert.equal(await accountName.getText(), '2nd account')
-      await delay(regularDelayMs)
-    })
-  })
-
-  describe('Switch back to original account', () => {
-    it('chooses the original account from the account menu', async () => {
-      await driver.findElement(By.css('.account-menu__icon')).click()
-      await delay(regularDelayMs)
-
-      const [originalAccountMenuItem] = await findElements(driver, By.css('.account-menu__name'))
-      await originalAccountMenuItem.click()
+    it('balance renders', async () => {
+      const balance = await findElement(driver, By.css('.transaction-view-balance__primary-balance'))
+      await driver.wait(until.elementTextMatches(balance, /100\s*ETH/))
       await delay(regularDelayMs)
     })
   })
@@ -291,12 +320,16 @@ describe('Using MetaMask with an existing account', function () {
       await inputAddress.sendKeys('0x2f318C334780961FB129D2a6c30D0763d9a5C970')
       await inputAmount.sendKeys('1')
 
+      const inputValue = await inputAmount.getAttribute('value')
+      assert.equal(inputValue, '1')
+
       // Set the gas limit
       const configureGas = await findElement(driver, By.css('.send-v2__gas-fee-display button'))
       await configureGas.click()
       await delay(regularDelayMs)
 
       const gasModal = await driver.findElement(By.css('span .modal'))
+
       const save = await findElement(driver, By.xpath(`//button[contains(text(), 'Save')]`))
       await save.click()
       await driver.wait(until.stalenessOf(gasModal))
@@ -311,70 +344,17 @@ describe('Using MetaMask with an existing account', function () {
     it('confirms the transaction', async function () {
       const confirmButton = await findElement(driver, By.xpath(`//button[contains(text(), 'Confirm')]`))
       await confirmButton.click()
-      await delay(regularDelayMs)
+      await delay(largeDelayMs)
     })
 
     it('finds the transaction in the transactions list', async function () {
       const transactions = await findElements(driver, By.css('.transaction-list-item'))
       assert.equal(transactions.length, 1)
 
-      const txValues = await findElements(driver, By.css('.transaction-list-item__amount--primary'))
-      assert.equal(txValues.length, 1)
-      assert.ok(/-1\s*ETH/.test(await txValues[0].getText()))
-    })
-  })
-
-  describe('Imports an account with private key', () => {
-    it('choose Create Account from the account menu', async () => {
-      await driver.findElement(By.css('.account-menu__icon')).click()
-      await delay(regularDelayMs)
-
-      const [importAccount] = await findElements(driver, By.xpath(`//div[contains(text(), 'Import Account')]`))
-      await importAccount.click()
-      await delay(regularDelayMs)
-    })
-
-    it('enter private key', async () => {
-      const privateKeyInput = await findElement(driver, By.css('#private-key-box'))
-      await privateKeyInput.sendKeys(testPrivateKey2)
-      await delay(regularDelayMs)
-      const importButtons = await findElements(driver, By.xpath(`//button[contains(text(), 'Import')]`))
-      await importButtons[0].click()
-      await delay(regularDelayMs)
-    })
-
-    it('should show the correct account name', async () => {
-      const [accountName] = await findElements(driver, By.css('.account-name'))
-      assert.equal(await accountName.getText(), 'Account 3')
-      await delay(regularDelayMs)
-    })
-
-    it('should show the imported label', async () => {
-      const [importedLabel] = await findElements(driver, By.css('.wallet-view__keyring-label'))
-      assert.equal(await importedLabel.getText(), 'IMPORTED')
-      await delay(regularDelayMs)
-    })
-  })
-
-  describe('Connects to a Hardware wallet', () => {
-    it('choose Connect Hardware Wallet from the account menu', async () => {
-      await driver.findElement(By.css('.account-menu__icon')).click()
-      await delay(regularDelayMs)
-
-      const [connectAccount] = await findElements(driver, By.xpath(`//div[contains(text(), 'Connect Hardware Wallet')]`))
-      await connectAccount.click()
-      await delay(regularDelayMs)
-    })
-
-    it('should open the TREZOR Connect popup', async () => {
-      const trezorButton = await findElements(driver, By.css('.hw-connect__btn'))
-      await trezorButton[1].click()
-      await delay(regularDelayMs)
-      const connectButtons = await findElements(driver, By.xpath(`//button[contains(text(), 'Connect')]`))
-      await connectButtons[0].click()
-      await delay(regularDelayMs)
-      const allWindows = await driver.getAllWindowHandles()
-      assert.equal(allWindows.length, 2)
+      if (process.env.SELENIUM_BROWSER !== 'firefox') {
+        const txValues = await findElement(driver, By.css('.transaction-list-item__amount--primary'))
+        await driver.wait(until.elementTextMatches(txValues, /-1\s*ETH/), 10000)
+      }
     })
   })
 })