+++ ui/app/components/transaction-view-balance/transaction-view-balance.component.js
@@ -37,13 +37,13 @@ export default class TransactionViewBalance extends PureComponent {
             className="transaction-view-balance__primary-balance"
             value={balance}
             type={PRIMARY}
-            ethNumberOfDecimals={3}
+            ethNumberOfDecimals={4}
           />
           <UserPreferencedCurrencyDisplay
             className="transaction-view-balance__secondary-balance"
             value={balance}
             type={SECONDARY}
-            ethNumberOfDecimals={3}
+            ethNumberOfDecimals={4}
           />
         </div>
       )